refactor(dot11): extract header geometry helpers into header module
